add PerseusError to perseus-core
