+++ packages/perseus/src/components/graphie.tsx
@@ -1,10 +1,11 @@
+import {Errors} from "@khanacademy/perseus-core";
 import $ from "jquery";
 import * as React from "react";
 import ReactDOM from "react-dom";
 import _ from "underscore";
 
 import InteractiveUtil from "../interactive2/interactive-util";
-import {Errors, Log} from "../logging/log";
+import {Log} from "../logging/log";
 import Util from "../util";
 import GraphUtils from "../util/graph-utils";
 import {Graphie as GraphieDrawingContext} from "../util/graphie";
