+++ packages/perseus/src/util/interactive.ts
@@ -24,8 +24,8 @@ import "../jquery.mobile.vmouse";
 import {Arrowhead} from "../interactive2/arrowhead";
 import WrappedEllipse from "../interactive2/wrapped-ellipse";
 import WrappedLine from "../interactive2/wrapped-line";
-import {Errors} from "../logging/log";
-import {PerseusError} from "../perseus-error";
+
+import {Errors, PerseusError} from "@khanacademy/perseus-error";
 
 import KhanColors from "./colors";
 import {clockwise, reverseVector} from "./geometry";
